dynamic client review form functionality is created

+++ src/Pages/MyBookings/MyBookings.jsx
@@ -3,17 +3,19 @@ import { AuthContext } from "../../AuthProvider/AuthProvider";
 import axios from "axios";
 import useAxiosSecure from "../../Hooks/useAxiosSecure";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 // import MyBooking from "./MyBooking";
 
 
 const MyBookings = () => {
-    const { user,updateRoom } = useContext(AuthContext);
+    const { user, updateRoom } = useContext(AuthContext);
     const [bookings, setBookings] = useState([]);
     const axiosSecure = useAxiosSecure();
 
     const userEmail = user.email;
-    
+
     // console.log(userEmail)
+   
     useEffect(() => {
         // fetch(`http://localhost:5000/my-bookings/${userEmail}`{credentials:"include"})
         //     .then(res => res.json())
@@ -21,46 +23,62 @@ const MyBookings = () => {
         //         // console.log(data)
         //         setBookings(data)
         //     })
-
-
-        axiosSecure.get(`/my-bookings/${userEmail}`,{withCredentials: true})
-        .then(res=>{
-            console.log(res.data)
-            setBookings(res.data)
-        })
-
-
-    }, [userEmail,axiosSecure]);
-
-
-    const handleDelete = (id) => {
-        console.log(3333, id);
         
 
-        fetch(`http://localhost:5000/booking-delete/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                const remaining = bookings.filter((booking) => booking._id !== id)
-                setBookings(remaining)
+        axiosSecure.get(`/my-bookings/${userEmail}`, { withCredentials: true })
+            .then(res => {
+                console.log(333333333333,res.data);
+                
+                setBookings(res.data)
+                
             })
 
-        //jjjjjjjjjjjjjj
-        const updateRoomInfo = { availability: true }
 
-        fetch(`http://localhost:5000/rooms-upadate/${updateRoom}`, {
-            updateRoomInfo,
-            method: 'PATCH',
-            headers: { 'content-type': 'application/json' },
-            body: JSON.stringify(updateRoomInfo)
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
+    }, [userEmail, axiosSecure]);
+   
 
-            })
+    const handleDelete = (_id,id) => {
+        console.log(3333, id);
+
+        Swal.fire({
+            title: "Are you sure to delete booking ?",
+           
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "confirm"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/booking-delete/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        const remaining = bookings.filter((booking) => booking._id !== _id)
+                        setBookings(remaining)
+                    })
+
+                //jjjjjjjjjjjjjj
+                const updateRoomInfo = { availability: true }
+
+                fetch(`http://localhost:5000/rooms-upadate/${id}`, {
+                    updateRoomInfo,
+                    method: 'PATCH',
+                    headers: { 'content-type': 'application/json' },
+                    body: JSON.stringify(updateRoomInfo)
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+
+                    })
+
+            }
+        });
+
+
 
 
 
@@ -93,7 +111,7 @@ const MyBookings = () => {
                                 bookings.map((booking, index) => <tr key={index}>
                                     <th>
                                         <label>
-                                            <span onClick={() => handleDelete(booking._id)} className="btn">X</span>
+                                            <span onClick={() => handleDelete(booking._id ,booking.id)} className="btn">X</span>
                                         </label>
                                     </th>
                                     <td>
@@ -122,7 +140,7 @@ const MyBookings = () => {
                                         <button className="btn btn-ghost btn-xs">confirm</button>
                                     </th>
                                     <th>
-                                        <Link to='/user-comment'><button className="btn btn-ghost btn-xs">Add Review</button></Link>
+                                        <Link to={`/user-comment/${booking.id}`}><button className="btn btn-ghost btn-xs">Add Review</button></Link>
                                     </th>
                                 </tr>)
                             }
